Refactor stage builder
